add MovieAlreadyExists exception

+++ Backend/src/main/java/com/fmi/cinema/cinema/repository/TicketRepository.java
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TicketRepository extends JpaRepository<Ticket, Integer> {
-
+public interface TicketRepository extends JpaRepository<Ticket, Long> {
     List<Ticket> findAll();
 }
